Extract shared install/update pipeline out of install.rs
